pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orderly-trader"
version = "0.1.0"
description = "LLM-powered trading strategy for Orderly Network perpetual futures"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "websockets>=12.0",
    "orderly-evm-connector>=0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
